pyproject.toml
[project]
name = "cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
